Create initial skeleton of tagit vcs
utils: register tagit (*) as possible vcs;
tagit: add dummy functions of vcs interface.

(*) Tagit is meant to be a simple wrapper over the git, which is based on tags
and simulates the versioning of software as a git repository. Major versions
(if specified by certain parameter) correspond to checked-out branches
and minor versions corresponds to tagged commits.

This should allow to use Perun as a simple store, without the automation,
and the need for having a supported vcs.